requirements.txt
alembic==1.18.4
blinker==1.9.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
pillow==12.1.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reportlab==4.4.10
six==1.17.0
SQLAlchemy==2.0.47
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
